refactor: namespace Cypher-specific Query types under Application\Query\Cypher
Move QueryService, QueryRequest, QueryResult, and QueryResultNormalizer
into a new Application\Query\Cypher namespace and drop their language
prefix from the class names. Rename the factory accordingly:
NeoWikiExtension::newQueryService() becomes newCypherQueryService().

Per ADR 19 each backend is its own plugin with its own native query
language. Putting Cypher-specific types in their own subnamespace makes
that grouping explicit and gives a clean parallel slot for the future
Application\Query\Sparql sibling.

Things that stay generic (no namespace move):
- QueryLimits — timeout and row cap apply to any backend.
- The QueryException hierarchy — most subclasses (EmptyQueryException,
  WriteQueryRejectedException, ParameterMissingException,
  QueryTimeoutException, BackendUnavailableException, InternalQueryException)
  are language-agnostic. CypherSyntaxException stays in the shared
  namespace because it carries its language in the class name.

git mv preserves history on the six relocated files.
