+++ src/Application/Query/Cypher/QueryResultNormalizer.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace ProfessionalWiki\NeoWiki\Application\Query;
+namespace ProfessionalWiki\NeoWiki\Application\Query\Cypher;
 
 use Laudis\Neo4j\Types\CypherList;
 use Laudis\Neo4j\Types\CypherMap;
