+++ src/Application/Query/Cypher/QueryService.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace ProfessionalWiki\NeoWiki\Application\Query;
+namespace ProfessionalWiki\NeoWiki\Application\Query\Cypher;
 
 use Laudis\Neo4j\Databags\SummarizedResult;
 use Laudis\Neo4j\Exception\Neo4jException;
