+++ src/EntryPoints/REST/QueryCypherApi.php
@@ -16,8 +16,8 @@ use ProfessionalWiki\NeoWiki\Application\Query\Exception\QueryException;
 use ProfessionalWiki\NeoWiki\Application\Query\Exception\QueryTimeoutException;
 use ProfessionalWiki\NeoWiki\Application\Query\Exception\WriteQueryRejectedException;
 use ProfessionalWiki\NeoWiki\Application\Query\QueryLimits;
-use ProfessionalWiki\NeoWiki\Application\Query\QueryRequest;
-use ProfessionalWiki\NeoWiki\Application\Query\QueryService;
+use ProfessionalWiki\NeoWiki\Application\Query\Cypher\QueryRequest;
+use ProfessionalWiki\NeoWiki\Application\Query\Cypher\QueryService;
 use Throwable;
 use Wikimedia\ParamValidator\ParamValidator;
