+++ src/EntryPoints/Scribunto/CypherQueryRunner.php
@@ -7,8 +7,8 @@ namespace ProfessionalWiki\NeoWiki\EntryPoints\Scribunto;
 use MediaWiki\Context\RequestContext;
 use ProfessionalWiki\NeoWiki\Application\Query\Exception\QueryException;
 use ProfessionalWiki\NeoWiki\Application\Query\QueryLimits;
-use ProfessionalWiki\NeoWiki\Application\Query\QueryRequest;
-use ProfessionalWiki\NeoWiki\Application\Query\QueryService;
+use ProfessionalWiki\NeoWiki\Application\Query\Cypher\QueryRequest;
+use ProfessionalWiki\NeoWiki\Application\Query\Cypher\QueryService;
 use RuntimeException;
 
 class CypherQueryRunner {
